requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
anthropic
email-validator